migration_source/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
cryptography>=41.0.0
python-multipart>=0.0.6
pyjwt>=2.6.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pytest>=7.3.1
httpx>=0.24.0
websockets>=10.4
starlette>=0.27.0
jinja2>=3.1.2
aiofiles>=23.1.0
itsdangerous>=2.1.2
redis>=4.5.5
email-validator>=2.0.0
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.3
pandas>=2.0.2
